Added check to allow user delete only their own listings

+++ api/controllers/listings/deleteListing.controller.js
@@ -7,7 +7,15 @@ const handleDeleteListing = async (req, res) => {
 
   if (!id) throw Error("Listing ID is required");
 
-  await Listing.findByIdAndDelete(id);
+  const listingDetails = await Listing.findById(id);
+
+  if (!listing) throw Error("Listing does not exist");
+
+  //Check if listing belongs to current logged in user
+  if (listingDetails.owner_id !== req.user._id)
+    throw Error("You can only delete your own listings");
+
+  const listing = await Listing.findByIdAndDelete(id);
 
   res.status(200).json({
     message: "Listing deleted successfully",
